Implemented Ceph Mounting
Super buggy, I've implemented the cephadore mounting and basic movement too. I've started rewriting the rigidFPS script so that it switches camera from the FPS to the Ceph's Mounted Camera, so that the player can look around whilst on the ceph whilst not being able to move their Player character. Super basic, no animation, nor sounds. Press F to mount the ceph :) - unable to turn around because of the camera bug, can move on the vertical axis but not the horizontal? idk wtf is going on whether its parametres are frozen or what.. ultimately that means you can't rotate the direction of the ceph in a typical way that you would steer something..